fix ci test scripts

+++ src/DataTypes/Serializations/SerializationInfoObjectToFetch.cpp
@@ -108,7 +108,7 @@ void SerializationInfoObjectToFetch::fromJSON(const Poco::JSON::Object & object)
             "Expected: {}, got: {}", elems.size(), subcolumns->size());
 
     for (size_t i = 0; i < elems.size(); ++i)
-        elems[i]->fromJSON(*subcolumns->getObject(i));
+        elems[i]->fromJSON(*subcolumns->getObject(static_cast<uint32_t>(i)));
 }
 
 }
