+++ src/Interpreters/ExpressionAnalyzer.cpp
@@ -771,7 +771,7 @@ bool ExpressionAnalyzer::makeVectorScanDescriptions(ActionsDAGPtr & actions)
         }
 
         /// top_k is get from limit N
-        vector_scan_desc.topk = syntax->limit_length;
+        vector_scan_desc.topk = static_cast<int>(syntax->limit_length);
 
         LOG_DEBUG(log, "[makeVectorScanDescriptions] create vector scan function: {}", node->name);
