+++ src/Storages/StorageReplicatedMergeTree.cpp
@@ -9470,7 +9470,7 @@ void StorageReplicatedMergeTree::updateVectorIndexInfoZookeeper()
 
         try
         {
-            UInt32 queue_size = getSettings()->max_queue_size_to_consider_replica_as_synced;
+            size_t queue_size = getSettings()->max_queue_size_to_consider_replica_as_synced;
             LOG_INFO(log, "Wait for replica syncing, target queue size: {}", queue_size);
 
             watch.start();
